requirements.txt
streamlit>=1.28.0
pandas>=2.0.0
numpy>=1.24.0
scikit-learn>=1.3.0
plotly>=5.14.0
matplotlib>=3.7.0
seaborn>=0.12.0
snowflake-connector-python>=3.0.0
Pillow>=10.0.0
python-dotenv>=1.0.0
